# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public intake and authenticated administration for Nexo Growth.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PublicConfig {
  /** @nullable */
  contactEmail: string | null;
  /** @nullable */
  gaMeasurementId: string | null;
}

export type SubmissionStatus = typeof SubmissionStatus[keyof typeof SubmissionStatus];


export const SubmissionStatus = {
  NEW: 'NEW',
  CONTACTED: 'CONTACTED',
  QUALIFIED: 'QUALIFIED',
  MATCHED: 'MATCHED',
  CLOSED: 'CLOSED',
  REJECTED: 'REJECTED',
} as const;

export type SubmissionType = typeof SubmissionType[keyof typeof SubmissionType];


export const SubmissionType = {
  BUYER: 'BUYER',
  PROVIDER: 'PROVIDER',
} as const;

export type BuyerSubmissionInputTimeline = typeof BuyerSubmissionInputTimeline[keyof typeof BuyerSubmissionInputTimeline];


export const BuyerSubmissionInputTimeline = {
  Inmediatamente: 'Inmediatamente',
  '1–3_meses': '1–3 meses',
  '3–6_meses': '3–6 meses',
  Más_de_6_meses: 'Más de 6 meses',
} as const;

export type BuyerSubmissionInputBudget = typeof BuyerSubmissionInputBudget[keyof typeof BuyerSubmissionInputBudget];


export const BuyerSubmissionInputBudget = {
  Menos_de_COP_$50M: 'Menos de COP $50M',
  'COP_$50M–$100M': 'COP $50M–$100M',
  'COP_$100M–$500M': 'COP $100M–$500M',
  'COP_$500M–$1000M': 'COP $500M–$1.000M',
  Más_de_COP_$1000M: 'Más de COP $1.000M',
  No_definido: 'No definido',
} as const;

export interface BuyerSubmissionInput {
  /**
     * @minLength 2
     * @maxLength 140
     */
  name: string;
  /**
     * @minLength 2
     * @maxLength 140
     */
  jobTitle: string;
  /**
     * @minLength 2
     * @maxLength 180
     */
  company: string;
  /** @maxLength 500 */
  website?: string;
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 6
     * @maxLength 60
     */
  phone: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  industry: string;
  /**
     * @minLength 2
     * @maxLength 100
     */
  country: string;
  /** @maxLength 120 */
  city?: string;
  /**
     * @minLength 2
     * @maxLength 240
     */
  productService: string;
  /**
     * @minLength 10
     * @maxLength 5000
     */
  description: string;
  timeline: BuyerSubmissionInputTimeline;
  budget: BuyerSubmissionInputBudget;
  /** @maxLength 240 */
  source?: string;
  consent: true;
  /** @maxLength 500 */
  honeypot?: string;
}

export type ProviderSubmissionInputSuccessFeeInterest = typeof ProviderSubmissionInputSuccessFeeInterest[keyof typeof ProviderSubmissionInputSuccessFeeInterest];


export const ProviderSubmissionInputSuccessFeeInterest = {
  Sí: 'Sí',
  Quiero_conocer_el_modelo: 'Quiero conocer el modelo',
  No: 'No',
} as const;

export interface ProviderSubmissionInput {
  /**
     * @minLength 2
     * @maxLength 140
     */
  name: string;
  /**
     * @minLength 2
     * @maxLength 140
     */
  jobTitle: string;
  /**
     * @minLength 2
     * @maxLength 180
     */
  company: string;
  /** @maxLength 500 */
  website?: string;
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 6
     * @maxLength 60
     */
  phone: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  industry: string;
  /**
     * @minLength 2
     * @maxLength 100
     */
  country: string;
  /**
     * @minLength 2
     * @maxLength 500
     */
  operatingRegions: string;
  /**
     * @minLength 2
     * @maxLength 240
     */
  productService: string;
  /**
     * @minLength 10
     * @maxLength 3000
     */
  problemSolved: string;
  /**
     * @minLength 2
     * @maxLength 500
     */
  targetIndustries: string;
  /**
     * @minLength 2
     * @maxLength 180
     */
  customerSize: string;
  /**
     * @minLength 2
     * @maxLength 180
     */
  averageTicket: string;
  /**
     * @minLength 2
     * @maxLength 1000
     */
  idealCustomer: string;
  successFeeInterest: ProviderSubmissionInputSuccessFeeInterest;
  consent: true;
  /** @maxLength 500 */
  honeypot?: string;
}

export interface SubmissionReceipt {
  id: number;
  type: SubmissionType;
  status: SubmissionStatus;
  createdAt: string;
}

export interface AdminLoginInput {
  /**
     * @minLength 1
     * @maxLength 512
     */
  password: string;
}

export interface AdminSessionStatus {
  authenticated: boolean;
}

export interface SubmissionStatusUpdate {
  status: SubmissionStatus;
}

export interface AdminSubmission {
  id: number;
  type: SubmissionType;
  name: string;
  jobTitle: string;
  company: string;
  /** @nullable */
  website?: string | null;
  email: string;
  phone: string;
  industry: string;
  country: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  operatingRegions?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  timeline?: string | null;
  /** @nullable */
  budget?: string | null;
  /** @nullable */
  productService?: string | null;
  /** @nullable */
  problemSolved?: string | null;
  /** @nullable */
  targetIndustries?: string | null;
  /** @nullable */
  customerSize?: string | null;
  /** @nullable */
  averageTicket?: string | null;
  /** @nullable */
  idealCustomer?: string | null;
  /** @nullable */
  successFeeInterest?: string | null;
  /** @nullable */
  source?: string | null;
  status: SubmissionStatus;
  createdAt: string;
}

export type AdminSubmissions = AdminSubmission[];

export interface AdminSummary {
  total: number;
  buyers: number;
  providers: number;
  latest: AdminSubmission[];
}

export type ListAdminSubmissionsParams = {
type?: ListAdminSubmissionsType;
status?: SubmissionStatus;
};

export type ListAdminSubmissionsType = typeof ListAdminSubmissionsType[keyof typeof ListAdminSubmissionsType];


export const ListAdminSubmissionsType = {
  BUYER: 'BUYER',
  PROVIDER: 'PROVIDER',
} as const;