# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Public intake and authenticated administration for Nexo Growth.
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: public
    description: Public site configuration and intake
  - name: admin
    description: Authenticated submission management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /public-config:
    get:
      operationId: getPublicConfig
      tags: [public]
      summary: Get safe public website configuration
      responses:
        "200":
          description: Public contact and analytics configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicConfig"
  /submissions/buyer:
    post:
      operationId: submitBuyer
      tags: [public]
      summary: Submit a buyer opportunity request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyerSubmissionInput"
      responses:
        "201":
          description: Submission received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionReceipt"
        "400":
          description: Invalid submission
        "409":
          description: A recent duplicate submission was detected
        "429":
          description: Rate limit exceeded
  /submissions/provider:
    post:
      operationId: submitProvider
      tags: [public]
      summary: Submit a provider company profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderSubmissionInput"
      responses:
        "201":
          description: Submission received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionReceipt"
        "400":
          description: Invalid submission
        "409":
          description: A recent duplicate submission was detected
        "429":
          description: Rate limit exceeded
  /admin/session:
    get:
      operationId: getAdminSession
      tags: [admin]
      summary: Check whether the current browser has an admin session
      responses:
        "200":
          description: Session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSessionStatus"
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Authenticate an administrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSessionStatus"
        "400":
          description: Invalid request
        "401":
          description: Authentication failed
        "429":
          description: Rate limit exceeded
  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: End the current administrator session
      responses:
        "204":
          description: Session ended
  /admin/summary:
    get:
      operationId: getAdminSummary
      tags: [admin]
      summary: Get submission totals and the most recent requests
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSummary"
        "401":
          description: Authentication required
  /admin/submissions:
    get:
      operationId: listAdminSubmissions
      tags: [admin]
      summary: List submissions for the administration dashboard
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [BUYER, PROVIDER]
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/SubmissionStatus"
      responses:
        "200":
          description: Submission list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSubmissions"
        "401":
          description: Authentication required
  /admin/submissions/{id}:
    patch:
      operationId: updateAdminSubmission
      tags: [admin]
      summary: Update a submission status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionStatusUpdate"
      responses:
        "200":
          description: Updated submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSubmission"
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "404":
          description: Submission not found
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    PublicConfig:
      type: object
      properties:
        contactEmail:
          type: ["string", "null"]
        gaMeasurementId:
          type: ["string", "null"]
      required: [contactEmail, gaMeasurementId]
    SubmissionStatus:
      type: string
      enum: [NEW, CONTACTED, QUALIFIED, MATCHED, CLOSED, REJECTED]
    SubmissionType:
      type: string
      enum: [BUYER, PROVIDER]
    BuyerSubmissionInput:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 140 }
        jobTitle: { type: string, minLength: 2, maxLength: 140 }
        company: { type: string, minLength: 2, maxLength: 180 }
        website: { type: string, maxLength: 500 }
        email: { type: string, format: email, maxLength: 254 }
        phone: { type: string, minLength: 6, maxLength: 60 }
        industry: { type: string, minLength: 2, maxLength: 120 }
        country: { type: string, minLength: 2, maxLength: 100 }
        city: { type: string, maxLength: 120 }
        productService: { type: string, minLength: 2, maxLength: 240 }
        description: { type: string, minLength: 10, maxLength: 5000 }
        timeline:
          type: string
          enum: ["Inmediatamente", "1–3 meses", "3–6 meses", "Más de 6 meses"]
        budget:
          type: string
          enum:
            - "Menos de COP $50M"
            - "COP $50M–$100M"
            - "COP $100M–$500M"
            - "COP $500M–$1.000M"
            - "Más de COP $1.000M"
            - "No definido"
        source: { type: string, maxLength: 240 }
        consent: { type: boolean, const: true }
        honeypot: { type: string, maxLength: 500 }
      required:
        - name
        - jobTitle
        - company
        - email
        - phone
        - industry
        - country
        - productService
        - description
        - timeline
        - budget
        - consent
    ProviderSubmissionInput:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 140 }
        jobTitle: { type: string, minLength: 2, maxLength: 140 }
        company: { type: string, minLength: 2, maxLength: 180 }
        website: { type: string, maxLength: 500 }
        email: { type: string, format: email, maxLength: 254 }
        phone: { type: string, minLength: 6, maxLength: 60 }
        industry: { type: string, minLength: 2, maxLength: 120 }
        country: { type: string, minLength: 2, maxLength: 100 }
        operatingRegions: { type: string, minLength: 2, maxLength: 500 }
        productService: { type: string, minLength: 2, maxLength: 240 }
        problemSolved: { type: string, minLength: 10, maxLength: 3000 }
        targetIndustries: { type: string, minLength: 2, maxLength: 500 }
        customerSize: { type: string, minLength: 2, maxLength: 180 }
        averageTicket: { type: string, minLength: 2, maxLength: 180 }
        idealCustomer: { type: string, minLength: 2, maxLength: 1000 }
        successFeeInterest:
          type: string
          enum: ["Sí", "Quiero conocer el modelo", "No"]
        consent: { type: boolean, const: true }
        honeypot: { type: string, maxLength: 500 }
      required:
        - name
        - jobTitle
        - company
        - email
        - phone
        - industry
        - country
        - operatingRegions
        - productService
        - problemSolved
        - targetIndustries
        - customerSize
        - averageTicket
        - idealCustomer
        - successFeeInterest
        - consent
    SubmissionReceipt:
      type: object
      properties:
        id: { type: integer }
        type: { $ref: "#/components/schemas/SubmissionType" }
        status: { $ref: "#/components/schemas/SubmissionStatus" }
        createdAt: { type: string, format: date-time }
      required: [id, type, status, createdAt]
    AdminLoginInput:
      type: object
      properties:
        password: { type: string, minLength: 1, maxLength: 512 }
      required: [password]
    AdminSessionStatus:
      type: object
      properties:
        authenticated: { type: boolean }
      required: [authenticated]
    SubmissionStatusUpdate:
      type: object
      properties:
        status: { $ref: "#/components/schemas/SubmissionStatus" }
      required: [status]
    AdminSubmission:
      type: object
      properties:
        id: { type: integer }
        type: { $ref: "#/components/schemas/SubmissionType" }
        name: { type: string }
        jobTitle: { type: string }
        company: { type: string }
        website: { type: ["string", "null"] }
        email: { type: string }
        phone: { type: string }
        industry: { type: string }
        country: { type: string }
        city: { type: ["string", "null"] }
        operatingRegions: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        timeline: { type: ["string", "null"] }
        budget: { type: ["string", "null"] }
        productService: { type: ["string", "null"] }
        problemSolved: { type: ["string", "null"] }
        targetIndustries: { type: ["string", "null"] }
        customerSize: { type: ["string", "null"] }
        averageTicket: { type: ["string", "null"] }
        idealCustomer: { type: ["string", "null"] }
        successFeeInterest: { type: ["string", "null"] }
        source: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/SubmissionStatus" }
        createdAt: { type: string, format: date-time }
      required:
        - id
        - type
        - name
        - jobTitle
        - company
        - email
        - phone
        - industry
        - country
        - status
        - createdAt
    AdminSubmissions:
      type: array
      items:
        $ref: "#/components/schemas/AdminSubmission"
    AdminSummary:
      type: object
      properties:
        total: { type: integer }
        buyers: { type: integer }
        providers: { type: integer }
        latest:
          type: array
          items:
            $ref: "#/components/schemas/AdminSubmission"
      required: [total, buyers, providers, latest]