# lib/api-zod/src/generated/types/adminLoginInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public intake and authenticated administration for Nexo Growth.
 * OpenAPI spec version: 0.1.0
 */

export interface AdminLoginInput {
  /**
     * @minLength 1
     * @maxLength 512
     */
  password: string;
}