# lib/api-zod/src/generated/types/adminSessionStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public intake and authenticated administration for Nexo Growth.
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSessionStatus {
  authenticated: boolean;
}