# lib/api-zod/src/generated/types/adminSubmission.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public intake and authenticated administration for Nexo Growth.
 * OpenAPI spec version: 0.1.0
 */
import type { SubmissionStatus } from './submissionStatus';
import type { SubmissionType } from './submissionType';

export interface AdminSubmission {
  id: number;
  type: SubmissionType;
  name: string;
  jobTitle: string;
  company: string;
  /** @nullable */
  website?: string | null;
  email: string;
  phone: string;
  industry: string;
  country: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  operatingRegions?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  timeline?: string | null;
  /** @nullable */
  budget?: string | null;
  /** @nullable */
  productService?: string | null;
  /** @nullable */
  problemSolved?: string | null;
  /** @nullable */
  targetIndustries?: string | null;
  /** @nullable */
  customerSize?: string | null;
  /** @nullable */
  averageTicket?: string | null;
  /** @nullable */
  idealCustomer?: string | null;
  /** @nullable */
  successFeeInterest?: string | null;
  /** @nullable */
  source?: string | null;
  status: SubmissionStatus;
  createdAt: Date;
}