# lib/api-zod/src/generated/types/buyerSubmissionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public intake and authenticated administration for Nexo Growth.
 * OpenAPI spec version: 0.1.0
 */
import type { BuyerSubmissionInputBudget } from './buyerSubmissionInputBudget';
import type { BuyerSubmissionInputTimeline } from './buyerSubmissionInputTimeline';

export interface BuyerSubmissionInput {
  /**
     * @minLength 2
     * @maxLength 140
     */
  name: string;
  /**
     * @minLength 2
     * @maxLength 140
     */
  jobTitle: string;
  /**
     * @minLength 2
     * @maxLength 180
     */
  company: string;
  /** @maxLength 500 */
  website?: string;
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 6
     * @maxLength 60
     */
  phone: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  industry: string;
  /**
     * @minLength 2
     * @maxLength 100
     */
  country: string;
  /** @maxLength 120 */
  city?: string;
  /**
     * @minLength 2
     * @maxLength 240
     */
  productService: string;
  /**
     * @minLength 10
     * @maxLength 5000
     */
  description: string;
  timeline: BuyerSubmissionInputTimeline;
  budget: BuyerSubmissionInputBudget;
  /** @maxLength 240 */
  source?: string;
  consent: true;
  /** @maxLength 500 */
  honeypot?: string;
}