# lib/api-zod/src/generated/types/submissionType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public intake and authenticated administration for Nexo Growth.
 * OpenAPI spec version: 0.1.0
 */

export type SubmissionType = typeof SubmissionType[keyof typeof SubmissionType];


export const SubmissionType = {
  BUYER: 'BUYER',
  PROVIDER: 'PROVIDER',
} as const;